Completed assignment
I get RSpec errors on the ArgumentErrors. I was expecting those tests
to pass since I told RSpec to expect those errors, but maybe that’s
just how it works. It’s a little irritating to see red text when things
go according to plan, but I’ll get used to it.